+++ src/components/feature/myProfile/MyprofileMain.tsx
@@ -1,11 +1,11 @@
 import { Tab, TabGroup, TabList, TabPanel, TabPanels } from '@headlessui/react';
 
-import MyKeyboardList from '@/components/feature/myProfile/MyKeyboardList';
+import MyKeyboardArea from '@/components/feature/myProfile/MyKeyboardArea';
 import MyReviewList from '@/components/feature/myProfile/MyReviewList';
 
 const MYPAGE_MENUS = [
   { title: '내가 쓴 후기', component: MyReviewList },
-  { title: '내가 등록한 키보드', component: MyKeyboardList },
+  { title: '내가 등록한 키보드', component: MyKeyboardArea },
 ];
 
 const MyprofileMain = () => {
